+++ core/r4300/r4300.h
@@ -40,11 +40,11 @@
 #include <core/r4300/Plugin.hpp>
 
 extern precomp_instr* PC;
-extern unsigned long vr_op;
+extern uint32_t vr_op;
 
 extern precomp_block *blocks[0x100000], *actual;
 extern void (*interp_ops[64])(void);
-extern int fast_memory;
+extern int32_t fast_memory;
 extern std::unique_ptr<Plugin> video_plugin;
 extern std::unique_ptr<Plugin> audio_plugin;
 extern std::unique_ptr<Plugin> input_plugin;
@@ -54,32 +54,32 @@ extern volatile bool emu_launched;
 extern volatile bool emu_paused;
 extern volatile bool core_executing;
 extern size_t g_total_frames;
-extern int stop, llbit;
-extern long long int reg[32], hi, lo;
-extern long long int local_rs, local_rt;
-extern unsigned long reg_cop0[32];
-extern long local_rs32, local_rt32;
-extern unsigned long jump_target;
+extern int32_t stop, llbit;
+extern int64_t reg[32], hi, lo;
+extern int64_t local_rs, local_rt;
+extern uint32_t reg_cop0[32];
+extern int32_t local_rs32, local_rt32;
+extern uint32_t jump_target;
 extern double* reg_cop1_double[32];
 extern float* reg_cop1_simple[32];
-extern long reg_cop1_fgr_32[32];
-extern long long int reg_cop1_fgr_64[32];
-extern long FCR0, FCR31;
+extern int32_t reg_cop1_fgr_32[32];
+extern int64_t reg_cop1_fgr_64[32];
+extern int32_t FCR0, FCR31;
 extern tlb tlb_e[32];
-extern unsigned long delay_slot, skip_jump, dyna_interp;
-extern unsigned long long int debug_count;
-extern unsigned long dynacore;
-extern unsigned long interpcore;
-extern unsigned int next_interrupt, CIC_Chip;
-extern int rounding_mode, trunc_mode, round_mode, ceil_mode, floor_mode;
-extern short x87_status_word;
-extern unsigned long last_addr, interp_addr;
+extern uint32_t delay_slot, skip_jump, dyna_interp;
+extern uint64_t debug_count;
+extern uint32_t dynacore;
+extern uint32_t interpcore;
+extern uint32_t next_interrupt, CIC_Chip;
+extern int32_t rounding_mode, trunc_mode, round_mode, ceil_mode, floor_mode;
+extern int16_t x87_status_word;
+extern uint32_t last_addr, interp_addr;
 extern char invalid_code[0x100000];
-extern unsigned long jump_to_address;
+extern uint32_t jump_to_address;
 extern std::atomic<bool> screen_invalidated;
-extern int vi_field;
-extern unsigned long next_vi;
-extern int compare_core_mode;
+extern int32_t vi_field;
+extern uint32_t next_vi;
+extern int32_t compare_core_mode;
 extern bool g_vr_fast_forward;
 extern bool g_vr_frame_skipped;
 
@@ -173,7 +173,7 @@ void pure_interpreter();
 void compare_core();
 extern void jump_to_func();
 void update_count();
-int check_cop1_unusable();
+int32_t check_cop1_unusable();
 void terminate_emu();
 
 #define jump_to(a) { jump_to_address = a; jump_to_func(); }
@@ -190,8 +190,8 @@ void terminate_emu();
 #define VR_SECTION_LUA_ATVI 3
 
 #ifdef VR_PROFILE
-void start_section(int section_type);
-void end_section(int section_type);
+void start_section(int32_t section_type);
+void end_section(int32_t section_type);
 #else
 #define start_section(x)
 #define end_section(x)
