+++ core/r4300/regimm.cpp
@@ -51,7 +51,7 @@ void BLTZ()
 void BLTZ_OUT()
 {
     local_rs = core_irs;
-    jump_target = (long)PC->f.i.immediate;
+    jump_target = (int32_t)PC->f.i.immediate;
     PC++;
     delay_slot = 1;
     PC->ops();
@@ -65,7 +65,7 @@ void BLTZ_OUT()
 
 void BLTZ_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs < 0)
     {
         update_count();
@@ -94,7 +94,7 @@ void BGEZ()
 void BGEZ_OUT()
 {
     local_rs = core_irs;
-    jump_target = (long)PC->f.i.immediate;
+    jump_target = (int32_t)PC->f.i.immediate;
     PC++;
     delay_slot = 1;
     PC->ops();
@@ -108,7 +108,7 @@ void BGEZ_OUT()
 
 void BGEZ_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs >= 0)
     {
         update_count();
@@ -142,7 +142,7 @@ void BLTZL_OUT()
 {
     if (core_irs < 0)
     {
-        jump_target = (long)PC->f.i.immediate;
+        jump_target = (int32_t)PC->f.i.immediate;
         PC++;
         delay_slot = 1;
         PC->ops();
@@ -159,7 +159,7 @@ void BLTZL_OUT()
 
 void BLTZL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs < 0)
     {
         update_count();
@@ -193,7 +193,7 @@ void BGEZL_OUT()
 {
     if (core_irs >= 0)
     {
-        jump_target = (long)PC->f.i.immediate;
+        jump_target = (int32_t)PC->f.i.immediate;
         PC++;
         delay_slot = 1;
         PC->ops();
@@ -210,7 +210,7 @@ void BGEZL_OUT()
 
 void BGEZL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs >= 0)
     {
         update_count();
@@ -247,7 +247,7 @@ void BLTZAL_OUT()
     reg[31] = PC->addr + 8;
     if ((&core_irs) != (reg + 31))
     {
-        jump_target = (long)PC->f.i.immediate;
+        jump_target = (int32_t)PC->f.i.immediate;
         PC++;
         delay_slot = 1;
         PC->ops();
@@ -263,7 +263,7 @@ void BLTZAL_OUT()
 
 void BLTZAL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs < 0)
     {
         update_count();
@@ -300,7 +300,7 @@ void BGEZAL_OUT()
     reg[31] = PC->addr + 8;
     if ((&core_irs) != (reg + 31))
     {
-        jump_target = (long)PC->f.i.immediate;
+        jump_target = (int32_t)PC->f.i.immediate;
         PC++;
         delay_slot = 1;
         PC->ops();
@@ -316,7 +316,7 @@ void BGEZAL_OUT()
 
 void BGEZAL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs >= 0)
     {
         update_count();
@@ -360,7 +360,7 @@ void BLTZALL_OUT()
     {
         if (local_rs < 0)
         {
-            jump_target = (long)PC->f.i.immediate;
+            jump_target = (int32_t)PC->f.i.immediate;
             PC++;
             delay_slot = 1;
             PC->ops();
@@ -379,7 +379,7 @@ void BLTZALL_OUT()
 
 void BLTZALL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs < 0)
     {
         update_count();
@@ -423,7 +423,7 @@ void BGEZALL_OUT()
     {
         if (local_rs >= 0)
         {
-            jump_target = (long)PC->f.i.immediate;
+            jump_target = (int32_t)PC->f.i.immediate;
             PC++;
             delay_slot = 1;
             PC->ops();
@@ -442,7 +442,7 @@ void BGEZALL_OUT()
 
 void BGEZALL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if (core_irs >= 0)
     {
         update_count();
