+++ core/r4300/x86/gcop1_l.cpp
@@ -36,12 +36,12 @@
 void gencvt_s_l()
 {
 #ifdef INTERPRET_CVT_S_L
-	gencallinterp((unsigned long)CVT_S_L, 0);
+	gencallinterp((uint32_t)CVT_S_L, 0);
 #else
     gencheck_cop1_unusable();
-    mov_eax_memoffs32((unsigned long*)(&reg_cop1_double[dst->f.cf.fs]));
+    mov_eax_memoffs32((uint32_t*)(&reg_cop1_double[dst->f.cf.fs]));
     fild_preg32_qword(EAX);
-    mov_eax_memoffs32((unsigned long*)(&reg_cop1_simple[dst->f.cf.fd]));
+    mov_eax_memoffs32((uint32_t*)(&reg_cop1_simple[dst->f.cf.fd]));
     fstp_preg32_dword(EAX);
 #endif
 }
@@ -49,12 +49,12 @@ void gencvt_s_l()
 void gencvt_d_l()
 {
 #ifdef INTERPRET_CVT_D_L
-	gencallinterp((unsigned long)CVT_D_L, 0);
+	gencallinterp((uint32_t)CVT_D_L, 0);
 #else
     gencheck_cop1_unusable();
-    mov_eax_memoffs32((unsigned long*)(&reg_cop1_double[dst->f.cf.fs]));
+    mov_eax_memoffs32((uint32_t*)(&reg_cop1_double[dst->f.cf.fs]));
     fild_preg32_qword(EAX);
-    mov_eax_memoffs32((unsigned long*)(&reg_cop1_double[dst->f.cf.fd]));
+    mov_eax_memoffs32((uint32_t*)(&reg_cop1_double[dst->f.cf.fd]));
     fstp_preg32_qword(EAX);
 #endif
 }
